requirements.txt
flask
flask_restful
pillow
requests
numpy
